refactor: split models.py into domain modules within blowcomotion app
