--- blowcomotion/models.py
@@ -1,2127 +1,0 @@
-import datetime
-import uuid as uuid_module
-
-from modelcluster.contrib.taggit import ClusterTaggableManager
-from modelcluster.fields import ParentalKey, ParentalManyToManyField
-from modelcluster.models import ClusterableModel
-from taggit.models import ItemBase, TagBase
-from wagtail import blocks
-from wagtail.admin.panels import FieldPanel, MultiFieldPanel, MultipleChooserPanel
-from wagtail.contrib.settings.models import BaseSiteSetting, register_setting
-from wagtail.documents import get_document_model
-from wagtail.fields import RichTextField, StreamField
-from wagtail.images.models import AbstractImage, AbstractRendition, Image
-from wagtail.models import (
-    DraftStateMixin,
-    LockableMixin,
-    Orderable,
-    Page,
-    RevisionMixin,
-)
-from wagtail.search import index
-
-from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.core.validators import MinValueValidator
-from django.db import models
-
-from blowcomotion import blocks as blowcomotion_blocks
-from members.utils import validate_birthday
-
-
-def get_default_expiration_date():
-    return datetime.date.today() + datetime.timedelta(days=1)
-
-@register_setting
-class NotificationBanner(BaseSiteSetting):
-    message = RichTextField(blank=True, null=True)
-    expiration_date = models.DateField(
-        blank=True,
-        null=True,
-        default=get_default_expiration_date,
-        help_text="Date when the banner will no longer be displayed. Leave blank for no expiration.",
-    )
-
-
-@register_setting
-class SiteSettings(BaseSiteSetting):
-    logo = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-    )
-    favicon = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-        help_text="Image used for the browser favicon. Upload a square image for best results.",
-    )
-    footer_text = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text="Text to display in the footer",
-    )
-    email = models.EmailField(blank=True, null=True)
-    instagram = models.URLField(blank=True, null=True)
-    facebook = models.URLField(blank=True, null=True)
-    header_menus = StreamField(
-        [
-            ("menu_item", blowcomotion_blocks.MenuItem()),
-        ],
-        blank=True,
-        null=True,
-    )
-    contact_form_email_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive contact form submissions",
-    )
-    join_band_form_email_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive join band form submissions",
-    )
-    booking_form_email_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive booking form submissions",
-    )
-    feedback_form_email_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive feedback form submissions",
-    )
-    donate_form_email_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive donate form submissions",
-    )
-    birthday_summary_email_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive monthly birthday summary emails",
-    )
-    instrument_rental_notification_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive instrument rental notifications",
-    )
-    instrument_rental_policy = RichTextField(
-        blank=True,
-        help_text="Lending policy text displayed on the instrument rental request form.",
-    )
-    venmo_donate_url = models.URLField(
-        blank=True,
-        null=True,
-        help_text="URL to Venmo donation page",
-    )
-    square_donate_url = models.URLField(
-        blank=True,
-        null=True,
-        help_text="URL to Square donation page",
-    )
-    patreon_url = models.URLField(
-        blank=True,
-        null=True,
-        help_text="URL to Patreon page",
-    )
-    
-    attendance_cleanup_days = models.IntegerField(
-        default=90,
-        help_text=(
-            "Number of days since last seeing a member before they are marked inactive "
-            "(attendance cleanup) and before instrument renters receive a nag email."
-        ),
-    )
-    nag_cooldown_days = models.IntegerField(
-        default=7,
-        help_text="Days to wait before sending another nag email to the same renter.",
-    )
-    attendance_report_notification_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive attendance report notifications",
-    )
-    member_signup_notification_recipients = models.CharField(
-        max_length=1024,
-        blank=True,
-        null=True,
-        help_text="Comma-separated list of email addresses to receive member signup notifications",
-    )
-    
-    panels = [
-        MultiFieldPanel([
-            FieldPanel('logo'),
-            FieldPanel('favicon'),
-            FieldPanel('footer_text'),
-            FieldPanel('email'),
-        ], heading="Site Branding"),
-        
-        MultiFieldPanel([
-            FieldPanel('instagram'),
-            FieldPanel('facebook'),
-        ], heading="Social Media"),
-        
-        FieldPanel('header_menus'),
-        
-        MultiFieldPanel([
-            FieldPanel('contact_form_email_recipients'),
-            FieldPanel('join_band_form_email_recipients'),
-            FieldPanel('booking_form_email_recipients'),
-            FieldPanel('feedback_form_email_recipients'),
-            FieldPanel('donate_form_email_recipients'),
-            FieldPanel('birthday_summary_email_recipients'),
-            FieldPanel('instrument_rental_notification_recipients'),
-            FieldPanel('attendance_report_notification_recipients'),
-            FieldPanel('member_signup_notification_recipients'),
-        ], heading="Form Email Recipients"),
-
-        MultiFieldPanel([
-            FieldPanel('instrument_rental_policy'),
-        ], heading="Instrument Rental Policy"),
-
-        MultiFieldPanel([
-            FieldPanel('venmo_donate_url'),
-            FieldPanel('square_donate_url'),
-            FieldPanel('patreon_url'),
-        ], heading="Donation Links"),
-        
-        MultiFieldPanel([
-            FieldPanel('attendance_cleanup_days'),
-            FieldPanel('nag_cooldown_days'),
-        ], heading="Attendance Cleanup Notifications", help_text="Configure attendance cleanup settings."),
-    ]
-
-    class Meta:
-        permissions = [
-            ("access_dev_tools", "Can access developer data dump tools"),
-            ("access_real_data_exports", "Can access real member data dumps and CSV exports"),
-        ]
-
-
-class CustomImage(AbstractImage):
-    # Add any extra fields to image here
-
-    # To add a caption field:
-    caption = models.CharField(max_length=255, blank=True)
-
-    admin_form_fields = Image.admin_form_fields + (
-        # Then add the field names here to make them appear in the form:
-        "caption",
-    )
-
-    @property
-    def default_alt_text(self):
-        # Force editors to add specific alt text if description is empty.
-        # Do not use image title which is typically derived from file name.
-        return getattr(self, "description", None)
-
-
-class CustomRendition(AbstractRendition):
-    image = models.ForeignKey(
-        CustomImage, on_delete=models.CASCADE, related_name="renditions"
-    )
-
-    class Meta:
-        unique_together = (("image", "filter_spec", "focal_point_key"),)
-
-
-class Chart(models.Model):
-    song = ParentalKey("blowcomotion.Song", related_name="charts")
-    pdf = models.ForeignKey(
-        get_document_model(),
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-    )
-    drive_pdf_url = models.URLField(null=True, blank=True)
-    part = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text=" e.g. '2nd Trombone' If left blank, instrument name will be used.",
-    )
-    instrument = models.ForeignKey("blowcomotion.Instrument", on_delete=models.CASCADE)
-    drive_file_id = models.CharField(max_length=255, null=True, blank=True, db_index=True)
-    drive_modified_time = models.DateTimeField(null=True, blank=True)
-    drive_imported_at = models.DateTimeField(null=True, blank=True)
-
-    # Index song and instrument fields for searchability in the chart library
-    search_fields = [
-        index.SearchField("song__title", partial_match=True, boost=2),
-        index.SearchField("instrument__name", partial_match=True),
-    ]
-
-    def clean(self):
-        if not self.pdf and not self.drive_pdf_url:
-            raise ValidationError("A chart must have either a PDF document or a Drive PDF URL.")
-
-    def __str__(self):
-        return f"{self.song.title} - {self.instrument.name} - {self.part}" if self.part else f"{self.song.title} - {self.instrument.name}"
-
-
-class SongConductor(Orderable):
-    song = ParentalKey("blowcomotion.Song", related_name="conductors")
-    member = models.ForeignKey("blowcomotion.Member", on_delete=models.CASCADE)
-
-
-class SongSoloist(Orderable):
-    song = ParentalKey("blowcomotion.Song", related_name="soloists")
-    member = models.ForeignKey("blowcomotion.Member", on_delete=models.CASCADE)
-
-
-class SongVideo(Orderable):
-    """Video URLs for a song (e.g., YouTube links to source performances)."""
-    song = ParentalKey("blowcomotion.Song", related_name="videos")
-    url = models.URLField(help_text="URL to video (YouTube, Vimeo, etc.)")
-    title = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text="Optional title/description for this video (e.g., 'Original by Rebirth Brass Band')"
-    )
-
-    panels = [
-        FieldPanel("url"),
-        FieldPanel("title"),
-    ]
-
-    def __str__(self):
-        return f"{self.song.title} - {self.title or self.url}"
-
-
-time_signature_choices = [
-    ("4/4", "4/4"),
-    ("3/4", "3/4"),
-    ("2/4", "2/4"),
-    ("2/2", "2/2"),
-    ("6/8", "6/8"),
-    ("12/8", "12/8"),
-    ("5/4", "5/4"),
-    ("7/8", "7/8"),
-    ("9/8", "9/8"),
-]
-
-key_signature_choices = [
-    ("C", "C"),
-    ("F", "F"),
-    ("Bb", "Bb"),
-    ("Eb", "Eb"),
-    ("Ab", "Ab"),
-    ("Db", "Db"),
-    ("Gb", "Gb"),
-    ("G", "G"),
-    ("D", "D"),
-    ("A", "A"),
-    ("E", "E"),
-    ("B", "B"),
-    ("F#", "F#"),
-    ("C#", "C#"),
-]
-
-
-tonality_choices = [
-    ("major", "Major"),
-    ("minor", "Minor"),
-    ("blues", "Blues"),
-]
-
-class Song(ClusterableModel, index.Indexed):
-    title = models.CharField(max_length=255)
-    time_signature = models.CharField(max_length=255, blank=True, null=True, choices=time_signature_choices)
-    key_signature = models.CharField(max_length=255, blank=True, null=True, choices=key_signature_choices)
-    tonality = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        choices=tonality_choices,
-    )
-    tempo = models.IntegerField(
-        blank=True,
-        null=True,
-        help_text="Tempo in BPM (Beats Per Minute)",
-    )
-    style = models.CharField(max_length=255, blank=True, null=True)
-    composer = models.CharField(max_length=255, blank=True, null=True)
-    arranger = models.CharField(max_length=255, blank=True, null=True)
-    form = models.TextField(blank=True, null=True, help_text="e.g. 'Intro, Verse, Chorus, Bridge, Outro or AABA'")
-    description = models.TextField(blank=True, null=True)
-    recording = models.ForeignKey(
-        "wagtailmedia.Media",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-    )
-    source_band = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text="e.g. 'Rebirth Brass Band', 'The Beatles', 'The Rolling Stones', etc.",
-    )
-    active = models.BooleanField(default=True)
-
-    search_fields = [
-        index.SearchField("title"),
-        index.AutocompleteField("title"),
-        index.SearchField("time_signature"),
-        index.SearchField("key_signature"),
-        index.SearchField("tonality"),
-        index.SearchField("arranger"),
-        index.SearchField("composer"),
-        index.SearchField("description"),
-        index.SearchField("style"),
-        index.SearchField("recording"),
-        index.SearchField("form"),
-        index.SearchField("tempo"),
-        index.SearchField("source_band"),
-        index.SearchField("active"),
-    ]
-
-    def __str__(self):
-        return self.title
-
-
-class EventSetlistSong(Orderable):
-    event = ParentalKey("blowcomotion.Event", related_name="setlist")
-    song = models.ForeignKey("blowcomotion.Song", on_delete=models.CASCADE)
-
-
-class Event(ClusterableModel, index.Indexed):
-    """
-    Model for events
-
-    Attributes:
-        title: CharField
-        date: DateField
-        time: TimeField
-        location: CharField
-        location_url: URLField
-        description: TextField
-        setlist: inline panel for songs
-    """
-
-    title = models.CharField(max_length=255)
-    date = models.DateField(blank=True, null=True)
-    time = models.TimeField(blank=True, null=True)
-    event_scroller_image = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-        help_text="Image to be used in the event scroller component",
-    )
-    location = models.CharField(
-        blank=True,
-        null=True,
-        max_length=255,
-        help_text="e.g. 'Mueller Lake Park, Austin, TX'",
-    )
-    location_url = models.URLField(
-        blank=True, null=True, help_text="URL for a map of the location"
-    )
-    description = models.TextField(blank=True, null=True)
-
-    search_fields = [
-        index.SearchField("title"),
-        index.SearchField("location"),
-        index.SearchField("description"),
-    ]
-
-    def __str__(self):
-        attributes = []
-        if self.date:
-            attributes.append(str(self.date))
-        if self.location:
-            attributes.append(self.location)
-        date_location = f" ({', '.join(attributes)})" if attributes else ""
-        return f"{self.title}{date_location}"
-
-
-class Section(ClusterableModel, index.Indexed):
-    name = models.CharField(max_length=255)
-
-    search_fields = [
-        index.SearchField("name"),
-    ]
-
-    def get_members(self):
-        """Active members whose primary or additional instrument belongs to this section."""
-        return Member.objects.filter(
-            is_active=True
-        ).filter(
-            models.Q(primary_instrument__section=self) |
-            models.Q(additional_instruments__instrument__section=self)
-        ).distinct()
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = "Section"
-        verbose_name_plural = "Sections"
-
-
-class SectionInstructor(Orderable):
-    section = ParentalKey("blowcomotion.Section", related_name="instructors")
-    instructor = models.ForeignKey("blowcomotion.Member", on_delete=models.CASCADE)
-
-    panels = [
-        "instructor",
-    ]
-
-
-class Instrument(models.Model, index.Indexed):
-    name = models.CharField(max_length=255)
-    description = models.TextField(blank=True, null=True)
-    section = models.ForeignKey(
-        "blowcomotion.Section",
-        null=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-    )
-    hide_from_rental = models.BooleanField(
-        default=False,
-        help_text=(
-            "Hide this instrument type from the rental request form. Use for instruments "
-            "we carry but don't offer for general rental — e.g. a bass clarinet added to "
-            "inventory because a member brought their own, but too rare to offer to others."
-        ),
-    )
-    hide_from_member_forms = models.BooleanField(
-        default=False,
-        help_text=(
-            "Hide this instrument type from member profile and signup instrument selectors. "
-            "Use for instruments members don't play but exist in inventory — e.g. a prop "
-            "instrument or one not assigned to any section."
-        ),
-    )
-    image = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-    )
-
-    search_fields = [
-        index.AutocompleteField("name"),
-        index.SearchField("name"),
-        index.SearchField("description"),
-    ]
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = "Instrument Type"
-        verbose_name_plural = "Instrument Types"
-
-
-class MemberInstrument(Orderable):
-    member = ParentalKey("blowcomotion.Member", related_name="additional_instruments")
-    instrument = models.ForeignKey("blowcomotion.Instrument", on_delete=models.CASCADE)
-
-    panels = [
-        "instrument",
-    ]
-
-
-class Member(RevisionMixin, ClusterableModel, index.Indexed):
-    """
-    Model for members of the organization
-
-    Attributes:
-        first_name: CharField
-        last_name: CharField
-        preferred_name: CharField
-        primary_instrument: ForeignKey - primary instrument (single choice)
-        additional_instruments: ManyToManyField - additional instruments through MemberInstrument
-        birth_month: IntegerField
-        birth_day: IntegerField
-        birth_year: IntegerField
-        join_date: DateField
-        is_active: BooleanField
-        bio: TextField
-        image: ForeignKey
-        instructor: BooleanField
-        board_member: BooleanField
-        renting: BooleanField
-        last_seen: DateField
-        separation_date: DateField
-        email: EmailField
-        gigomatic_username: CharField
-        gigomatic_id: IntegerField
-        phone: CharField
-        address: CharField
-        city: CharField
-        state: CharField
-        zip_code: CharField
-        country: CharField
-        notes: TextField
-        emergency_contact: TextField
-    """
-
-    first_name = models.CharField(max_length=255)
-    last_name = models.CharField(max_length=255)
-    preferred_name = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text="Name the member prefers to be called (optional)"
-    )
-    primary_instrument = models.ForeignKey(
-        "blowcomotion.Instrument",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="primary_members",
-        help_text="Primary instrument played by this member"
-    )
-    birth_month = models.IntegerField(
-        blank=True, 
-        null=True,
-        choices=[
-            (1, 'January'), (2, 'February'), (3, 'March'), (4, 'April'),
-            (5, 'May'), (6, 'June'), (7, 'July'), (8, 'August'),
-            (9, 'September'), (10, 'October'), (11, 'November'), (12, 'December')
-        ],
-        help_text="Birth month (1-12)"
-    )
-    birth_day = models.IntegerField(
-        blank=True, 
-        null=True,
-        help_text="Birth day (1-31)"
-    )
-    birth_year = models.IntegerField(
-        blank=True, 
-        null=True,
-        help_text="Birth year (YYYY)"
-    )
-    join_date = models.DateField(blank=True, null=True)
-    is_active = models.BooleanField(default=True)
-    bio = models.TextField(blank=True, null=True)
-    image = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-    )
-    instructor = models.BooleanField(default=False)
-    board_member = models.BooleanField(default=False)
-    renting = models.BooleanField(
-        default=False, 
-        help_text="Is the member renting an instrument? (Auto-updated when instruments are rented/returned)"
-    )
-    last_seen = models.DateField(blank=True, null=True, help_text="This field auto-populates whenever attendance is taken.")
-    separation_date = models.DateField(blank=True, null=True, help_text="Date of separation from the organization.")
-    reactivated_date = models.DateField(blank=True, null=True, help_text="Date when the member was reactivated (is_active changed to True).")
-    email = models.EmailField()
-    gigomatic_username = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text="Username used on Gig-O-Matic for this member",
-    )
-    gigomatic_id = models.IntegerField(
-        blank=True,
-        null=True,
-        help_text="ID used on Gig-O-Matic for this member",
-    )
-    phone = models.CharField(max_length=255, blank=True, null=True)
-    address = models.CharField(max_length=255, blank=True, null=True)
-    city = models.CharField(max_length=255, blank=True, null=True)
-    state = models.CharField(max_length=255, blank=True, null=True)
-    zip_code = models.CharField(max_length=255, blank=True, null=True)
-    country = models.CharField(max_length=255, blank=True, null=True)
-    notes = models.TextField(blank=True, null=True)
-    emergency_contact = models.TextField(
-        blank=True,
-        null=True,
-        help_text="Name and phone number of emergency contact",
-    )
-    inspired_by = models.TextField(
-        blank=True,
-        null=True,
-        help_text="What event or person inspired this member to join the band",
-    )
-
-    # Shirt size
-    SHIRT_SIZE_CHOICES = [
-        ('S', 'S'),
-        ('M', 'M'),
-        ('L', 'L'),
-        ('XL', 'XL'),
-        ('2XL', '2XL'),
-        ('3XL', '3XL'),
-        ('4XL', '4XL'),
-    ]
-    shirt_size = models.CharField(
-        max_length=10,
-        blank=True,
-        choices=SHIRT_SIZE_CHOICES,
-        help_text="Adult unisex shirt size",
-    )
-
-    # Dietary preferences (multi-select stored as JSON list)
-    dietary_preferences = models.JSONField(
-        blank=True,
-        default=list,
-        help_text="List of dietary preferences/practices",
-    )
-    dietary_other = models.TextField(
-        blank=True,
-        help_text="Other dietary preference details",
-    )
-
-    # Allergy information
-    has_allergies = models.BooleanField(
-        null=True,
-        blank=True,
-        help_text="Does the member have food/substance allergies?",
-    )
-    allergens = models.JSONField(
-        blank=True,
-        default=list,
-        help_text="List of allergens the member has",
-    )
-    allergens_other = models.TextField(
-        blank=True,
-        help_text="Other allergen details",
-    )
-    has_epipen = models.BooleanField(
-        null=True,
-        blank=True,
-        help_text="Does the member carry an Epi-Pen or other intervention?",
-    )
-    allergy_details = models.TextField(
-        blank=True,
-        help_text="Additional details about allergies or interventions",
-    )
-
-    # Confidential medical notes
-    medical_notes = models.TextField(
-        blank=True,
-        help_text="CONFIDENTIAL: Any medical concerns or allergies (not shared outside admin)",
-    )
-
-    # Auth
-    user = models.OneToOneField(
-        settings.AUTH_USER_MODEL,
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="member",
-    )
-    pending_email = models.EmailField(null=True, blank=True)
-
-    # Notification preferences
-    notify_rental_updates = models.BooleanField(default=True)
-    notify_reminders = models.BooleanField(default=True)
-    notify_announcements = models.BooleanField(default=True)
-
-    # Patreon cache — populated by validate_patreon_rentals / admin refresh
-    patreon_is_active = models.BooleanField(null=True, blank=True)
-    patreon_pledge_cents = models.PositiveIntegerField(null=True, blank=True)
-    patreon_last_charge_date = models.DateTimeField(null=True, blank=True)
-    patreon_last_charge_status = models.CharField(max_length=20, null=True, blank=True)
-    patreon_patron_since = models.DateTimeField(null=True, blank=True)
-    patreon_lifetime_cents = models.PositiveIntegerField(null=True, blank=True)
-    patreon_last_synced = models.DateTimeField(null=True, blank=True)
-
-    # Commenting out search_fields to avoid FTS indexing issues
-    # Admin search still works via snippet_viewsets.py search_fields
-    search_fields = [
-        index.SearchField("first_name"),
-        index.SearchField("last_name"),
-        index.SearchField("preferred_name"),
-        index.SearchField("gigomatic_username"),
-        index.AutocompleteField("first_name"),
-        index.AutocompleteField("last_name"),
-        index.AutocompleteField("preferred_name"),
-        index.AutocompleteField("gigomatic_username"),
-        index.SearchField("bio"),
-        index.SearchField("email"),
-        index.SearchField("phone"),
-        index.SearchField("address"),
-        index.SearchField("city"),
-        index.SearchField("state"),
-        index.SearchField("zip_code"),
-        index.SearchField("country"),
-        index.SearchField("notes"),
-    ]
-
-    def clean(self):
-        from django.core.exceptions import ValidationError
-
-        # Check for duplicate members based on first and last name
-        if self.first_name and self.last_name:
-            existing_members = Member.objects.filter(
-                first_name__iexact=self.first_name,
-                last_name__iexact=self.last_name
-            )
-            
-            # If this is an update (not a new member), exclude the current instance
-            if self.pk:
-                existing_members = existing_members.exclude(pk=self.pk)
-            
-            if existing_members.exists():
-                raise ValidationError(
-                    f"A member with the name '{self.first_name} {self.last_name}' already exists. "
-                    "Please check if this person is already in the system or use a different name."
-                )
-        
-        # Existing birthday validation
-        validate_birthday(self.birth_day, self.birth_month)
-
-    @property
-    def birthday(self):
-        """Return birthday as a date object if year is available, otherwise None"""
-        if self.birth_year and self.birth_month and self.birth_day:
-            import datetime
-            try:
-                return datetime.date(self.birth_year, self.birth_month, self.birth_day)
-            except ValueError:
-                return None
-        return None
-
-    @property
-    def birthday_display(self):
-        """Return a formatted birthday string for display"""
-        if self.birth_month and self.birth_day:
-            month_names = ['January', 'February', 'March', 'April', 'May', 'June',
-                         'July', 'August', 'September', 'October', 'November', 'December']
-            month_name = month_names[self.birth_month - 1]
-            if self.birth_year:
-                return f"{month_name} {self.birth_day}, {self.birth_year}"
-            else:
-                return f"{month_name} {self.birth_day}"
-        return None
-
-    @property
-    def full_name(self):
-        """Return the preferred full name for display."""
-        if self.preferred_name:
-            return f"{self.preferred_name} {self.last_name}"
-        return f"{self.first_name} {self.last_name}"
-
-    def get_gigo_id(self):
-        """
-        Get the Gig-O-Matic member ID for this member.
-        If gigomatic_id is not set, query the GO3 API by email to fetch it.
-        Returns the member ID or None if not found.
-        """
-        # Return cached value if already set
-        if self.gigomatic_id:
-            return self.gigomatic_id
-        
-        # Check if we have an email to query with
-        if not self.email:
-            return None
-        
-        # Import here to avoid circular imports
-        import logging
-        from urllib.parse import quote
-
-        from django.conf import settings
-
-        from gigs.gigo import make_gigo_api_request
-        
-        logger = logging.getLogger(__name__)
-        
-        # Check if API is configured
-        if not settings.GIGO_API_URL or not settings.GIGO_API_KEY:
-            logger.warning(f"GO3 API not configured, cannot query member ID for {self.email}")
-            return None
-        
-        # Query the member by email
-        try:
-            endpoint = f"/members/query?email={quote(self.email)}"
-            response = make_gigo_api_request(endpoint)
-            
-            if response and 'member_id' in response:
-                # Cache the member ID
-                self.gigomatic_id = response['member_id']
-                self.save(update_fields=['gigomatic_id'], sync_go3=False)
-                logger.info(f"Fetched and cached GO3 member ID {self.gigomatic_id} for {self.email}")
-                return self.gigomatic_id
-            else:
-                logger.warning(f"Could not find GO3 member ID for {self.email}")
-                return None
-        except Exception as e:
-            logger.error(f"Error querying GO3 API for member {self.email}: {e}")
-            return None
-    
-    def save(self, *args, **kwargs):
-        """
-        Override save to sync is_active status with Gig-O-Matic's is_occasional field.
-        When is_active changes:
-        - False (inactive) → occasional=True in GO3
-        - True (active) → occasional=False in GO3
-        
-        Args:
-            sync_go3 (bool): Whether to sync with GO3 API. Set to False for internal
-                            updates that don't need GO3 synchronization (default: True)
-        
-        Performance Note:
-            GO3 member verification is only performed when needed:
-            - When email changes
-            - When is_active changes
-            - When gigomatic_id or gigomatic_username are missing
-            This avoids unnecessary API calls on high-frequency updates like attendance.
-        """
-        import logging
-        from urllib.parse import quote
-
-        from django.conf import settings
-
-        from gigs.gigo import make_gigo_api_request
-        
-        logger = logging.getLogger(__name__)
-        
-        # Extract sync_go3 kwarg (defaults to True for backward compatibility)
-        sync_go3 = kwargs.pop('sync_go3', True)
-        
-        # Extract update_fields to check if specific fields are being updated
-        update_fields = kwargs.get('update_fields')
-        # Capture caller's value before any mutation (used by email drift guard below)
-        _caller_update_fields = update_fields
-
-        # Determine if this save operation can affect fields relevant for GO3 sync
-        sync_relevant_fields = (
-            update_fields is None
-            or 'email' in update_fields
-            or 'is_active' in update_fields
-        )
-        
-        # Track changes that require GO3 sync
-        is_active_changed = False
-        email_changed = False
-        old_is_active = None
-        old_email = None
-        
-        # Check if GO3 is configured to short-circuit unnecessary work
-        is_go3_configured = (settings.GIGO_API_URL and settings.GIGO_API_KEY)
-        
-        # Only fetch the old instance when GO3 sync is enabled, configured, and relevant fields may change
-        if self.pk and sync_go3 and is_go3_configured and sync_relevant_fields:
-            try:
-                old_instance = Member.objects.get(pk=self.pk)
-                old_is_active = old_instance.is_active
-                old_email = old_instance.email
-                
-                # Values are considered changed only when they actually differ
-                if old_is_active != self.is_active:
-                    is_active_changed = True
-                if old_email != self.email:
-                    email_changed = True
-            except Member.DoesNotExist:
-                pass
-        else:
-            # Even if sync_go3=False, check for reactivation to set reactivated_date
-            # Only fetch if is_active could be changing (avoid needless DB query)
-            if self.pk and (update_fields is None or 'is_active' in update_fields):
-                try:
-                    old_instance = Member.objects.get(pk=self.pk)
-                    old_is_active = old_instance.is_active
-                    if old_is_active != self.is_active:
-                        is_active_changed = True
-                except Member.DoesNotExist:
-                    pass
-        
-        # Set reactivated_date if transitioning from inactive to active
-        reactivated_date_set = False
-        if is_active_changed and old_is_active is not None and not old_is_active and self.is_active:
-            self.reactivated_date = datetime.date.today()
-            reactivated_date_set = True
-
-        # Ensure reactivated_date is persisted when using update_fields
-        if reactivated_date_set and 'update_fields' in kwargs and kwargs['update_fields'] is not None:
-            update_fields = kwargs['update_fields']
-            if isinstance(update_fields, (list, tuple, set)):
-                update_fields = set(update_fields)
-            else:
-                update_fields = {update_fields}
-            update_fields.add('reactivated_date')
-            kwargs['update_fields'] = update_fields
-        # Call parent save first
-        super().save(*args, **kwargs)
-
-        # Sync User.email and User.username if admin changed Member.email
-        if self.user_id and (_caller_update_fields is None or "email" in _caller_update_fields):
-            from django.contrib.auth import get_user_model as _get_user_model
-            _User = _get_user_model()
-            try:
-                _user = _User.objects.get(pk=self.user_id)
-                new_email = self.email or ""
-                if _user.email != new_email or _user.username != new_email:
-                    _user.email = new_email
-                    _user.username = new_email
-                    _user.save(update_fields=["email", "username"])
-                    logger.info(f"Synced User email/username for member {self.pk}")
-            except _User.DoesNotExist:
-                pass
-
-        # Only query GO3 when sync_go3=True AND sync_relevant_fields=True AND one of these conditions is met:
-        # 1. gigomatic_id or gigomatic_username is missing
-        # 2. Email changed
-        # 3. is_active changed
-        # This prevents unrelated field updates (e.g., update_fields=['renting']) from triggering API calls
-        should_verify_member = (
-            sync_go3 and 
-            self.email and 
-            is_go3_configured and
-            sync_relevant_fields and
-            (not self.gigomatic_id or not self.gigomatic_username or email_changed or is_active_changed)
-        )
-        
-        if should_verify_member:
-            try:
-                endpoint = f"/members/query?email={quote(self.email)}"
-                member_data = make_gigo_api_request(endpoint)
-                
-                if member_data and 'member_id' in member_data:
-                    # Track what needs updating
-                    update_fields = []
-                    
-                    # Update gigo_id if different or not set
-                    if self.gigomatic_id != member_data['member_id']:
-                        logger.info(f"Updating gigo_id for {self.full_name}: {self.gigomatic_id} → {member_data['member_id']}")
-                        self.gigomatic_id = member_data['member_id']
-                        update_fields.append('gigomatic_id')
-                    
-                    # Update username if different or not set
-                    if member_data.get('username') and self.gigomatic_username != member_data['username']:
-                        logger.info(f"Updating gigomatic_username for {self.full_name}: {self.gigomatic_username} → {member_data['username']}")
-                        self.gigomatic_username = member_data['username']
-                        update_fields.append('gigomatic_username')
-                    
-                    # Save updated fields if they changed
-                    if update_fields:
-                        # When using update_fields, we cannot use force_insert/force_update
-                        extra_save_kwargs = {
-                            key: value
-                            for key, value in kwargs.items()
-                            if key in ("using",)
-                        }
-                        extra_save_kwargs["update_fields"] = update_fields
-                        super().save(**extra_save_kwargs)
-                else:
-                    logger.warning(f"Could not verify member info from GO3 for {self.email}")
-            except Exception as e:
-                logger.warning(f"Error verifying member info from GO3: {e}")
-        
-        # If is_active changed, sync status with GO3
-        if is_active_changed and sync_go3:
-            # Check if API is configured before attempting status sync
-            if not is_go3_configured:
-                logger.debug(f"GO3 API not configured, skipping status sync for {self.full_name}")
-                return
-            
-            try:
-                # Get gigo_id (should be updated from verification above)
-                gigo_id = self.gigomatic_id or self.get_gigo_id()
-                if gigo_id:
-                    # Determine which band ID to use
-                    if settings.DEBUG:
-                        band_id = getattr(settings, 'GIGO_BAND_ID_LOCAL', None)
-                    else:
-                        band_id = getattr(settings, 'GIGO_BAND_ID', None)
-                    
-                    if band_id:
-                        # Determine desired occasional status based on is_active
-                        # inactive member (is_active=False) should be occasional (is_occasional=True)
-                        # active member (is_active=True) should be regular (is_occasional=False)
-                        desired_occasional = not self.is_active
-                        
-                        # Toggle member status in GO3
-                        endpoint = f"/bands/{band_id}/members/{gigo_id}/occasional"
-                        response = make_gigo_api_request(endpoint, method='PATCH')
-                        
-                        if response and 'is_occasional' in response:
-                            # Check if the result matches what we want
-                            if response['is_occasional'] == desired_occasional:
-                                # Success
-                                status = "occasional" if desired_occasional else "regular"
-                                logger.info(f"Synced member {self.full_name} to {status} in GO3")
-                            else:
-                                # Toggle didn't result in desired state, toggle again
-                                response2 = make_gigo_api_request(endpoint, method='PATCH')
-                                if response2 and response2.get('is_occasional') == desired_occasional:
-                                    status = "occasional" if desired_occasional else "regular"
-                                    logger.info(f"Synced member {self.full_name} to {status} in GO3 (after second toggle)")
-                                else:
-                                    logger.warning(f"Could not sync member {self.full_name} status to GO3")
-                        else:
-                            logger.warning(f"Could not sync member {self.full_name} status to GO3 - invalid response")
-                    else:
-                        logger.debug(f"GIGO_BAND_ID not configured, skipping GO3 sync for {self.full_name}")
-                else:
-                    logger.debug(f"No Gig-O-Matic ID found for member {self.full_name}, skipping GO3 sync")
-            except Exception as e:
-                # Log error but don't fail the save
-                logger.error(f"Error syncing member {self.full_name} status to GO3: {e}")
-    
-    def __str__(self):
-        return f"\"{self.preferred_name}\" {self.first_name} {self.last_name}" if self.preferred_name and self.preferred_name.strip().lower() != self.first_name.strip().lower() else f"{self.first_name} {self.last_name}"
-
-    def display_name(self):
-        # ponytail: __str__ unsortable in Wagtail admin (label_for_field returns builtin str as attr); this wrapper exposes admin_order_field
-        return str(self)
-    display_name.admin_order_field = 'first_name'
-    display_name.short_description = 'Name'
-
-
-class PasswordSetToken(models.Model):
-    member = models.ForeignKey(
-        Member, on_delete=models.CASCADE, related_name="set_password_tokens"
-    )
-    uuid = models.UUIDField(default=uuid_module.uuid4, unique=True, editable=False)
-    created_at = models.DateTimeField(auto_now_add=True)
-    used = models.BooleanField(default=False)
-    superseded = models.BooleanField(default=False)
-
-    def __str__(self):
-        return f"PasswordSetToken({self.member}, used={self.used})"
-
-
-class EmailChangeToken(models.Model):
-    member = models.ForeignKey(
-        Member, on_delete=models.CASCADE, related_name="email_change_tokens"
-    )
-    uuid = models.UUIDField(default=uuid_module.uuid4, unique=True, editable=False)
-    new_email = models.EmailField()
-    created_at = models.DateTimeField(auto_now_add=True)
-    used = models.BooleanField(default=False)
-
-    def __str__(self):
-        return f"EmailChangeToken({self.member} → {self.new_email})"
-
-
-class CachedGig(models.Model):
-    """
-    Model for caching gig data from the Gig-O-Matic API.
-    
-    Instead of making API calls on every page load, gig data is synced
-    periodically (via cron) and stored in the database for faster access.
-    
-    Attributes:
-        gig_id: The unique ID from Gig-O-Matic
-        title: Gig title
-        date: Date of the gig
-        time: Time of the gig (stored separately for querying)
-        address: Location/address of the gig
-        gig_status: Status from API (e.g., 'confirmed', 'unconfirmed')
-        band: Band name from the API
-        raw_data: Full JSON response from the API for this gig
-        last_synced: When this record was last updated from the API
-    """
-    gig_id = models.IntegerField(unique=True, db_index=True)
-    title = models.CharField(max_length=500)
-    date = models.DateField(db_index=True)
-    time = models.TimeField(null=True, blank=True)
-    address = models.TextField(blank=True, default='')
-    gig_status = models.CharField(max_length=50, db_index=True)
-    band = models.CharField(max_length=255, db_index=True)
-    raw_data = models.JSONField(default=dict)
-    last_synced = models.DateTimeField(auto_now=True)
-    
-    class Meta:
-        verbose_name = "Cached Gig"
-        verbose_name_plural = "Cached Gigs"
-        ordering = ['date', 'time']
-        indexes = [
-            models.Index(fields=['date', 'gig_status', 'band']),
-        ]
-    
-    def __str__(self):
-        return f"{self.title} ({self.date})"
-    
-    @classmethod
-    def get_gigs_for_date(cls, date_str, band=None, status='confirmed'):
-        """Get confirmed gigs for a specific date."""
-        from django.conf import settings
-        
-        band = band or getattr(settings, 'GIGO_BAND_NAME', 'Blowcomotion')
-        return cls.objects.filter(
-            date=date_str,
-            band__iexact=band,
-            gig_status__iexact=status,
-        )
-    
-    @classmethod
-    def get_upcoming_gigs(cls, band=None, status='confirmed'):
-        """Get all upcoming confirmed gigs from today onwards."""
-        from django.conf import settings
-        
-        band = band or getattr(settings, 'GIGO_BAND_NAME', 'Blowcomotion')
-        return cls.objects.filter(
-            date__gte=datetime.date.today(),
-            band__iexact=band,
-            gig_status__iexact=status,
-        ).order_by('date', 'time')
-    
-    @classmethod
-    def get_gig_by_id(cls, gig_id):
-        """Get a single gig by its Gig-O-Matic ID."""
-        try:
-            return cls.objects.get(gig_id=gig_id)
-        except cls.DoesNotExist:
-            return None
-    
-    def to_api_format(self):
-        """Return the gig data in the same format as the API response."""
-        return self.raw_data if self.raw_data else {
-            'id': self.gig_id,
-            'title': self.title,
-            'date': self.date.isoformat(),
-            'address': self.address,
-            'gig_status': self.gig_status,
-            'band': self.band,
-        }
-
-
-class AttendanceRecord(models.Model):
-    """
-    Model for tracking attendance at practice sessions
-    
-    Attributes:
-        date: DateField - date of practice
-        member: ForeignKey - reference to Member (nullable for guests)
-        guest_name: CharField - name of guest if not a member
-        notes: TextField - optional notes about attendance
-        created_at: DateTimeField - when record was created
-    """
-    
-    date = models.DateField(default=datetime.date.today)
-    member = models.ForeignKey(
-        "blowcomotion.Member",
-        null=True,
-        blank=True,
-        on_delete=models.CASCADE,
-        related_name="attendance_records"
-    )
-    played_instrument = models.ForeignKey(
-        "blowcomotion.Instrument",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="attendance_records_as_played",
-        help_text="Instrument the member played for this attendance entry"
-    )
-    guest_name = models.CharField(
-        max_length=255,
-        blank=True,
-        null=True,
-        help_text="Name of guest/visitor (leave blank for members)"
-    )
-    notes = models.TextField(blank=True, null=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    
-    class Meta:
-        unique_together = ['date', 'member']
-        ordering = ['-date', 'member__last_name']
-    
-    def clean(self):
-        if not self.member and not self.guest_name:
-            raise ValidationError("Either member or guest_name must be provided")
-        if self.member and self.guest_name:
-            raise ValidationError("Cannot specify both member and guest_name")
-    
-    def __str__(self):
-        if self.member:
-            instrument_label = f" ({self.played_instrument.name})" if self.played_instrument else ""
-            return f"{self.member}{instrument_label} - {self.date}"
-        else:
-            return f"{self.guest_name} (Guest) - {self.date}"
-
-
-class BasePage(Page):
-    class Meta:
-        abstract = True
-
-
-class BlankCanvasPage(BasePage):
-    template = "pages/blank_canvas_page.html"
-    sticky_content = StreamField(
-        [
-            ("column_layout", blowcomotion_blocks.ColumnLayoutBlock()),
-            ("hero", blowcomotion_blocks.HeroBlock()),
-            ("horizontal_rule", blowcomotion_blocks.HorizontalRuleBlock()),
-            ("image", blowcomotion_blocks.ImageBlock()),
-            ("rich_text", blowcomotion_blocks.AlignableRichtextBlock()),
-            ("adjustable_spacer", blowcomotion_blocks.AdjustableSpacerBlock()),
-            ("quote", blowcomotion_blocks.QuoteBlock()),
-        ],
-        blank=True,
-        null=True,
-        help_text="Content in this field will be fixed at the top of the page, above the main body content. Ideal for hero sections or important announcements that should always be visible."
-    )
-    body = StreamField(
-        [
-            ("accordion_list", blowcomotion_blocks.AccordionListBlock()),
-            ("booking_form", blowcomotion_blocks.BookingFormBlock(group="Forms")),
-            ("button", blowcomotion_blocks.ButtonBlock()),
-            ("chart_library", blowcomotion_blocks.ChartLibraryBlock()),
-            ("column_layout", blowcomotion_blocks.ColumnLayoutBlock()),
-            ("contact_form", blowcomotion_blocks.ContactFormBlock(group="Forms")),
-            ("countdown", blowcomotion_blocks.CountdownBlock()),
-            ("donate_form", blowcomotion_blocks.DonateFormBlock(group="Forms")),
-            ("events", blowcomotion_blocks.EventsBlock()),
-            ("full_width_image", blowcomotion_blocks.FullWidthImageBlock()),
-            ("hero", blowcomotion_blocks.HeroBlock()),
-            ("horizontal_rule", blowcomotion_blocks.HorizontalRuleBlock()),
-            ("image", blowcomotion_blocks.ImageBlock()),
-            ("image_carousel", blowcomotion_blocks.ImageCarouselBlock()),
-            ("join_band_form", blowcomotion_blocks.JoinBandFormBlock(group="Forms")),
-            ("member_signup_form", blowcomotion_blocks.MemberSignupFormBlock(group="Forms")),
-            ("jukebox", blowcomotion_blocks.JukeBoxBlock()),
-            ("multi_image_banner", blowcomotion_blocks.MultiImageBannerBlock()),
-            ("patreon_button", blowcomotion_blocks.PatreonButton()),
-            ("paypal_donate_button", blowcomotion_blocks.PayPalDonateButton()),
-            ("quoted_image", blowcomotion_blocks.QuotedImageBlock()),
-            ("rich_text", blowcomotion_blocks.AlignableRichtextBlock()),
-            ("adjustable_spacer", blowcomotion_blocks.AdjustableSpacerBlock()),
-            ("spacer", blowcomotion_blocks.SpacerBlock()),
-            ("square_donate_button", blowcomotion_blocks.SquareDonateButton()),
-            ("timeline", blowcomotion_blocks.TimelineBlock()),
-            ("upcoming_events", blowcomotion_blocks.UpcomingPublicGigs()),
-            ("venmo_donate_button", blowcomotion_blocks.VenmoDonateButton()),
-            ("video_feed", blowcomotion_blocks.VideoFeedBlock()),
-            ("quote", blowcomotion_blocks.QuoteBlock()),
-        ],
-        block_counts={
-            "hero": {"max_num": 1},
-        },
-        blank=True,
-        null=True,
-    )
-
-    content_panels = Page.content_panels + [
-        "sticky_content",
-        "body",
-    ]
-
-    def get_context(self, request):
-        context = super().get_context(request)
-        context["include_countdown_js"] = False
-        context["include_form_js"] = True # set to True for the feedback form
-        
-        if self.body:
-            has_notification_banner = NotificationBanner.for_request(request).message and (not NotificationBanner.for_request(request).expiration_date or NotificationBanner.for_request(request).expiration_date > datetime.date.today())
-            context["hero_header"] = self.body[0].block_type == "hero" and not has_notification_banner
-            context["bottom_countdown"] = self.body[-1].block_type == "countdown" and self.body[-1].value.get('countdown_date') and self.body[-1].value.get('countdown_date') > datetime.date.today()
-
-            for block in self.body:
-                if block.block_type == "countdown":
-                    context["include_countdown_js"] = True
-                # if block.block_type == "contact_form":
-                #     context["include_form_js"] = True
-                if context["include_form_js"] and context["include_countdown_js"]:
-                    break
-        else:
-            context["hero_header"] = False
-            context["bottom_countdown"] = False
-            context["include_quote_css"] = False
-            
-        return context
-    
-
-class WikiIndexPage(BlankCanvasPage):
-    """
-    Model for wiki index page
-
-    Attributes:
-        title: CharField
-        body: StreamField
-    """
-
-    template = "pages/blank_canvas_page.html"
-    subpage_types = ["blowcomotion.WikiPage"]
-    max_count = 1
-
-    class Meta:
-        verbose_name = "Wiki Index Page"
-        verbose_name_plural = "Wiki Index Pages"
-
-    def __str__(self):
-        return self.title
-    
-
-class WikiAuthor(Orderable):
-    """
-    Model for authors of wiki pages
-
-    Attributes:
-        page: ParentalKey
-        author: ForeignKey
-    """
-
-    page = ParentalKey("blowcomotion.WikiPage", related_name="authors")
-    author = models.ForeignKey("blowcomotion.Member", on_delete=models.CASCADE)
-
-    panels = [
-        FieldPanel("author"),
-    ]
-
-    def __str__(self):
-        return str(self.author)
-
-
-class WikiPage(BlankCanvasPage):
-    """
-    Model for wiki pages
-
-    Attributes:
-        title: CharField
-        body: StreamField
-    """
-
-    template = "pages/blank_canvas_page.html"
-    parent_page_types = ["blowcomotion.WikiIndexPage"]
-    subpage_types = ["blowcomotion.WikiPage"]
-
-    content_panels = [
-        MultipleChooserPanel("authors", chooser_field_name="author"),
-    ] + BlankCanvasPage.content_panels
-
-
-    class Meta:
-        verbose_name = "Wiki Page"
-        verbose_name_plural = "Wiki Pages"
-
-    def __str__(self):
-        return self.title
-
-
-class InstrumentStorageLocation(models.Model):
-    name = models.CharField(max_length=255)
-    description = models.TextField(blank=True, null=True)
-    street_address = models.CharField(max_length=255, blank=True, null=True)
-    city = models.CharField(max_length=255, blank=True, null=True)
-    state = models.CharField(max_length=255, blank=True, null=True)
-    zip_code = models.CharField(max_length=20, blank=True, null=True)
-    country = models.CharField(max_length=255, blank=True, null=True)
-    phone_number = models.CharField(max_length=50, blank=True, null=True)
-    email = models.EmailField(blank=True, null=True)
-    notes = models.TextField(blank=True, null=True)
-
-    def __str__(self):
-        return self.name
-
-
-class LibraryInstrument(DraftStateMixin, RevisionMixin, LockableMixin, ClusterableModel, index.Indexed):
-    """Track each physical instrument in the Blowcomotion inventory."""
-
-    STATUS_AVAILABLE = "available"
-    STATUS_RENTED = "rented"
-    STATUS_NEEDS_REPAIR = "needs_repair"
-    STATUS_OUT_FOR_REPAIR = "out_for_repair"
-    STATUS_DISPOSED = "disposed"
-
-    STATUS_CHOICES = [
-        (STATUS_AVAILABLE, "Available"),
-        (STATUS_RENTED, "Rented"),
-        (STATUS_NEEDS_REPAIR, "Needs Repair/Unplayable"),
-        (STATUS_OUT_FOR_REPAIR, "Out for Repair"),
-        (STATUS_DISPOSED, "Disposed"),
-    ]
-
-    instrument = models.ForeignKey(
-        "blowcomotion.Instrument",
-        on_delete=models.PROTECT,
-        related_name="library_inventory",
-        help_text="Instrument type (e.g. Trombone, Trumpet)",
-    )
-    status = models.CharField(
-        max_length=50,
-        choices=STATUS_CHOICES,
-        default=STATUS_AVAILABLE,
-        help_text="Current availability of this instrument",
-    )
-    serial_number = models.TextField(help_text="Serial number or other identifying marks")
-    hide_from_rental = models.BooleanField(
-        default=False,
-        help_text=(
-            "Hide this specific unit from the rental request form. Use when a unit is "
-            "damaged, reserved, or otherwise unavailable without removing it from inventory."
-        ),
-    )
-    hide_from_member_forms = models.BooleanField(
-        default=False,
-        help_text=(
-            "Hide this specific unit from member-facing instrument selectors. Use when "
-            "a unit should not be visible to members but remains in admin inventory."
-        ),
-    )
-    member = models.ForeignKey(
-        "blowcomotion.Member",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="rented_instruments",
-        help_text="Person currently storing or borrowing this instrument",
-    )
-    rental_date = models.DateField(
-        null=True,
-        blank=True,
-        help_text="Date the instrument was lent to the current member",
-    )
-    acquisition_cost = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Original cost of acquiring this instrument",
-    )
-    current_value = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Current estimated value of this instrument",
-    )
-    replacement_cost = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Estimated cost to replace this instrument",
-    )
-    patreon_active = models.BooleanField(
-        default=False,
-        help_text="Is the renter currently an active Patreon supporter?",
-    )
-    patreon_amount = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Monthly Patreon support amount",
-    )
-    last_nag_sent = models.DateField(
-        null=True,
-        blank=True,
-        help_text="Date the most recent nag email was sent to this renter.",
-    )
-    storage_location = models.ForeignKey(
-        "blowcomotion.InstrumentStorageLocation",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="stored_instruments",
-        help_text="Physical storage location for this instrument (e.g. organization locker). Leave blank if stored with member.",
-    )
-    comments = models.TextField(
-        blank=True,
-        null=True,
-        help_text="Additional context or maintenance notes",
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    search_fields = [
-        index.SearchField("serial_number"),
-        index.SearchField("instrument_name"),
-        index.SearchField("comments"),
-        index.AutocompleteField("serial_number"),
-    ]
-
-    class Meta:
-        ordering = ["instrument__name", "serial_number"]
-        verbose_name = "Library Instrument"
-        verbose_name_plural = "Library Instruments"
-
-    def __str__(self):
-        status_display = self.get_status_display()
-        if self.member:
-            return f"{self.instrument.name} ({self.serial_number[:20]}) - {status_display} - {self.member.full_name}"
-        return f"{self.instrument.name} ({self.serial_number[:20]}) - {status_display}"
-
-    @property
-    def instrument_name(self):
-        return self.instrument.name if self.instrument else ""
-
-    def clean(self):
-        super().clean()
-        if self.status == self.STATUS_RENTED and not self.member:
-            raise ValidationError({"member": "Rented instruments must be assigned to a member."})
-        # XOR member or storage_location must be set (can't have both or neither)
-        if self.member and self.storage_location:
-            raise ValidationError("An instrument cannot be stored with a member and in a storage location at the same time. Please choose one or the other.")
-        if not self.member and not self.storage_location:
-            raise ValidationError("An instrument must be stored with a member or in a storage location. Please choose one.")
-
-
-    def save(self, *args, **kwargs):
-        is_new = self.pk is None
-        old_status = None
-        old_member_id = None
-        old_member = None
-
-        if not is_new:
-            old_instance = LibraryInstrument.objects.filter(pk=self.pk).first()
-            if old_instance:
-                old_status = old_instance.status
-                old_member_id = old_instance.member_id
-                old_member = old_instance.member
-
-        super().save(*args, **kwargs)
-
-        # Update member renting status
-        members_to_update = set()
-        
-        # If this instrument is rented and has a member, ensure member.renting = True
-        if self.status == self.STATUS_RENTED and self.member:
-            members_to_update.add(self.member)
-            self.member.renting = True
-            self.member.save(update_fields=['renting'])
-        
-        # If member changed or status changed from rented, check old member's status
-        if old_member and old_member != self.member:
-            members_to_update.add(old_member)
-        
-        # Update old member's renting status if they have no other rentals
-        for member in members_to_update:
-            if member != self.member:  # Don't recheck current member
-                still_renting = LibraryInstrument.objects.filter(
-                    member=member,
-                    status=self.STATUS_RENTED
-                ).exists()
-                member.renting = still_renting
-                member.save(update_fields=['renting'])
-
-        if is_new:
-            InstrumentHistoryLog.objects.create(
-                library_instrument=self,
-                event_category=InstrumentHistoryLog.EVENT_ACQUISITION,
-                event_date=datetime.date.today(),
-                notes="Instrument added to library inventory",
-            )
-            if self.status == self.STATUS_RENTED:
-                InstrumentHistoryLog.objects.create(
-                    library_instrument=self,
-                    event_category=InstrumentHistoryLog.EVENT_OUT_FOR_LOAN,
-                    event_date=datetime.date.today(),
-                    notes=f"Instrument rented to {self.member.full_name}" if self.member else "Instrument marked as rented",
-                )
-
-        if old_status and old_status != self.status:
-            if old_status == self.STATUS_RENTED and self.status != self.STATUS_RENTED:
-                self.last_nag_sent = None
-                # Use update to avoid recursion; save() already called above via super()
-                LibraryInstrument.objects.filter(pk=self.pk).update(last_nag_sent=None)
-            self._create_status_change_log(old_status, self.status)
-        elif (
-            self.status == self.STATUS_RENTED
-            and old_member_id
-            and old_member_id != self.member_id
-        ):
-            # Member changed without updating status
-            InstrumentHistoryLog.objects.create(
-                library_instrument=self,
-                event_category=InstrumentHistoryLog.EVENT_RETURNED_FROM_LOAN,
-                event_date=datetime.date.today(),
-                notes=(
-                    f"Instrument returned from {old_member.full_name}"
-                    if old_member
-                    else "Instrument returned from loan"
-                ),
-            )
-            if self.member:
-                InstrumentHistoryLog.objects.create(
-                    library_instrument=self,
-                    event_category=InstrumentHistoryLog.EVENT_OUT_FOR_LOAN,
-                    event_date=datetime.date.today(),
-                    notes=f"Instrument rented to {self.member.full_name}",
-                )
-
-    def _create_status_change_log(self, old_status, new_status):
-        status_map = dict(self.STATUS_CHOICES)
-        notes = f"Status changed from {status_map.get(old_status)} to {status_map.get(new_status)}"
-        event_category = None
-
-        if new_status == self.STATUS_RENTED:
-            event_category = InstrumentHistoryLog.EVENT_OUT_FOR_LOAN
-            if self.member:
-                notes = f"Instrument rented to {self.member.full_name}"
-        elif new_status == self.STATUS_AVAILABLE:
-            if old_status == self.STATUS_RENTED:
-                event_category = InstrumentHistoryLog.EVENT_RETURNED_FROM_LOAN
-                notes = "Instrument returned from rental"
-            elif old_status == self.STATUS_OUT_FOR_REPAIR:
-                event_category = InstrumentHistoryLog.EVENT_RETURNED_FROM_REPAIR
-        elif new_status == self.STATUS_OUT_FOR_REPAIR:
-            event_category = InstrumentHistoryLog.EVENT_OUT_FOR_REPAIR
-            notes = "Instrument sent out for repair"
-        elif new_status == self.STATUS_NEEDS_REPAIR:
-            event_category = InstrumentHistoryLog.EVENT_LOCATION_CHECK
-            notes = "Instrument flagged as needing repair/unplayable"
-        elif new_status == self.STATUS_DISPOSED:
-            event_category = InstrumentHistoryLog.EVENT_DISPOSAL
-
-        if event_category:
-            InstrumentHistoryLog.objects.create(
-                library_instrument=self,
-                event_category=event_category,
-                event_date=datetime.date.today(),
-                notes=notes,
-            )
-
-    @property
-    def renter_inactive(self):
-        if self.status != self.STATUS_RENTED or not self.member or not self.member.last_seen:
-            return False
-        return (datetime.date.today() - self.member.last_seen).days >= 21 or not self.member.is_active
-
-
-class LibraryInstrumentPhoto(Orderable):
-    library_instrument = ParentalKey(
-        "blowcomotion.LibraryInstrument",
-        related_name="photos",
-        on_delete=models.CASCADE,
-    )
-    image = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        on_delete=models.CASCADE,
-        related_name="+",
-    )
-    caption = models.CharField(max_length=255, blank=True)
-
-    panels = [
-        FieldPanel("image"),
-        FieldPanel("caption"),
-    ]
-
-    def __str__(self):
-        return f"Photo for {self.library_instrument}"
-
-
-class InstrumentRentalNagLog(models.Model):
-    library_instrument = models.ForeignKey(
-        "blowcomotion.LibraryInstrument",
-        on_delete=models.CASCADE,
-        related_name="nag_logs",
-    )
-    member_name = models.CharField(max_length=255)
-    member_email = models.EmailField()
-    reasons = models.CharField(
-        max_length=255,
-        help_text='Comma-separated trigger reasons: "attendance", "patreon", or "attendance+patreon"',
-    )
-    sent_at = models.DateField()
-
-    class Meta:
-        ordering = ["-sent_at"]
-
-    def __str__(self):
-        return f"{self.member_name} — {self.sent_at} ({self.reasons})"
-
-
-class InstrumentHistoryLog(models.Model):
-    """
-    Model for tracking the event history of library instruments.
-    Auto-populated when status changes, but can also be manually created.
-    """
-    
-    # Event category choices
-    EVENT_ACQUISITION = 'acquisition'
-    EVENT_OUT_FOR_REPAIR = 'out_for_repair'
-    EVENT_RETURNED_FROM_REPAIR = 'returned_from_repair'
-    EVENT_OUT_FOR_LOAN = 'out_for_loan'
-    EVENT_RETURNED_FROM_LOAN = 'returned_from_loan'
-    EVENT_LOCATION_CHECK = 'location_check'
-    EVENT_DISPOSAL = 'disposal'
-    EVENT_RENTAL_NOTE = 'rental_note'
-    EVENT_RETURN_NOTE = 'return_note'
-    
-    EVENT_CHOICES = [
-        (EVENT_ACQUISITION, 'Acquisition'),
-        (EVENT_OUT_FOR_REPAIR, 'Out for Repair'),
-        (EVENT_RETURNED_FROM_REPAIR, 'Returned from Repair'),
-        (EVENT_OUT_FOR_LOAN, 'Out for Loan'),
-        (EVENT_RETURNED_FROM_LOAN, 'Returned from Loan'),
-        (EVENT_LOCATION_CHECK, 'Location Check'),
-        (EVENT_RENTAL_NOTE, 'Rental Note'),
-        (EVENT_RETURN_NOTE, 'Return Note'),
-        (EVENT_DISPOSAL, 'Disposal'),
-    ]
-    
-    library_instrument = ParentalKey(
-        'blowcomotion.LibraryInstrument',
-        on_delete=models.CASCADE,
-        related_name='history_logs'
-    )
-    event_date = models.DateField(
-        default=datetime.date.today,
-        help_text='Date of this event'
-    )
-    event_category = models.CharField(
-        max_length=50,
-        choices=EVENT_CHOICES,
-        help_text='Type of event'
-    )
-    notes = models.TextField(
-        blank=True,
-        help_text='Details about this event'
-    )
-    user = models.ForeignKey(
-        'auth.User',
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        help_text='User who created this log entry'
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    panels = [
-        FieldPanel('event_date'),
-        FieldPanel('event_category'),
-        FieldPanel('notes'),
-        FieldPanel('user'),
-    ]
-    
-    class Meta:
-        ordering = ['-event_date', '-created_at']
-        verbose_name = 'Instrument History Log'
-        verbose_name_plural = 'Instrument History Logs'
-    
-    def __str__(self):
-        return f"{self.library_instrument.instrument.name} - {self.get_event_category_display()} on {self.event_date}"
-
-
-class Equipment(ClusterableModel):
-    """Track non-instrument physical items in the storeroom (tables, canopy, signs, etc.)."""
-
-    STATUS_AVAILABLE = "available"
-    STATUS_NEEDS_REPAIR = "needs_repair"
-    STATUS_DISPOSED = "disposed"
-
-    STATUS_CHOICES = [
-        (STATUS_AVAILABLE, "Available"),
-        (STATUS_NEEDS_REPAIR, "Needs Repair"),
-        (STATUS_DISPOSED, "Disposed"),
-    ]
-
-    name = models.CharField(max_length=255)
-    serial_number = models.TextField(blank=True, help_text="Serial number or other identifying marks")
-    quantity = models.PositiveIntegerField(default=1)
-    status = models.CharField(
-        max_length=50,
-        choices=STATUS_CHOICES,
-        default=STATUS_AVAILABLE,
-    )
-    storage_location = models.ForeignKey(
-        "blowcomotion.InstrumentStorageLocation",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="equipment",
-        help_text="Where this item is stored (leave blank if location is unknown)",
-    )
-    acquisition_cost = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Original cost of acquiring this item",
-    )
-    current_value = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Current estimated value of this item",
-    )
-    replacement_cost = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        null=True,
-        blank=True,
-        validators=[MinValueValidator(0)],
-        help_text="Estimated cost to replace this item",
-    )
-    notes = models.TextField(blank=True, null=True, help_text="Additional context or maintenance notes")
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        ordering = ["name"]
-        verbose_name = "Equipment"
-        verbose_name_plural = "Equipment"
-
-
-class EquipmentPhoto(Orderable):
-    equipment = ParentalKey(
-        "blowcomotion.Equipment",
-        related_name="photos",
-        on_delete=models.CASCADE,
-    )
-    image = models.ForeignKey(
-        "blowcomotion.CustomImage",
-        on_delete=models.CASCADE,
-        related_name="+",
-    )
-    caption = models.CharField(max_length=255, blank=True)
-
-    panels = [
-        FieldPanel("image"),
-        FieldPanel("caption"),
-    ]
-
-    def __str__(self):
-        return f"Photo for {self.equipment}"
-
-
-class BaseFormSubmission(models.Model):
-    """
-    Base model for form submissions
-
-    Attributes:
-        name: CharField
-        email: EmailField
-        message: TextField
-        date_submitted: DateTimeField
-    This is an abstract model that can be inherited by other form submission models.
-    """
-
-    name = models.CharField(blank=True, null=True, max_length=255)
-    email = models.EmailField(blank=True, null=True, )
-    message = models.TextField(blank=True, null=True, )
-    date_submitted = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        abstract = True
-    
-
-class ContactFormSubmission(BaseFormSubmission):
-    """
-        Model for contact form submissions
-    """
-    newsletter_opt_in = models.BooleanField(
-        default=False,
-        help_text="Whether the user signed up for the newsletter",
-    )
-
-    def __str__(self):
-        return f"Contact Form Submission from {self.name} on {self.date_submitted}"
-    
-
-class FeedbackFormSubmission(BaseFormSubmission):
-    """
-        Model for feedback form submissions
-    """
-    submitted_from_page = models.CharField(
-        blank=True,
-        null=True,
-        max_length=255,
-        help_text="The URL of the page from which the feedback was submitted",
-    )
-
-    def __str__(self):
-        return f"Feedback Form Submission from {self.name} on {self.date_submitted}"
-
-
-class JoinBandFormSubmission(BaseFormSubmission):
-    """
-    Model for join band form submissions
-    """
-    instrument = models.CharField(
-        blank=True,
-        null=True,
-        max_length=255,
-        help_text="The instrument the person plays",
-    )
-    instrument_rental = models.CharField(
-        blank=True,
-        null=True,
-        max_length=10,
-        choices=[
-            ('yes', 'Yes, I would like to rent an instrument'),
-            ('no', 'No, I have my own instrument'),
-            ('maybe', 'I\'m not sure yet'),
-        ],
-        help_text="Whether the person wants to rent an instrument",
-    )
-    newsletter_opt_in = models.BooleanField(
-        default=False,
-        help_text="Whether the user signed up for the newsletter",
-    )
-
-    def __str__(self):
-        return f"Join Band Form Submission from {self.name} on {self.date_submitted}"
-
-
-class BookingFormSubmission(BaseFormSubmission):
-    """
-    Model for booking form submissions
-    """
-    event_date = models.DateField(
-        blank=True,
-        null=True,
-        help_text="Date of the event",
-    )
-    event_time = models.TimeField(
-        blank=True,
-        null=True,
-        help_text="Time of the event",
-    )
-    event_location = models.CharField(
-        blank=True,
-        null=True,
-        max_length=500,
-        help_text="Location of the event",
-    )
-    duration = models.CharField(
-        blank=True,
-        null=True,
-        max_length=255,
-        help_text="How long the band should play",
-    )
-    expected_guests = models.CharField(
-        blank=True,
-        null=True,
-        max_length=255,
-        help_text="Expected number of guests",
-    )
-    event_details = models.TextField(
-        blank=True,
-        null=True,
-        help_text="Specific event details and expectations",
-    )
-    budget = models.CharField(
-        blank=True,
-        null=True,
-        max_length=255,
-        help_text="Budget for the performance",
-    )
-    newsletter_opt_in = models.BooleanField(
-        default=False,
-        help_text="Whether the user signed up for the newsletter",
-    )
-
-    def __str__(self):
-        return f"Booking Form Submission from {self.name} on {self.date_submitted}"
-
-
-class DonateFormSubmission(BaseFormSubmission):
-    """
-    Model for donate form submissions
-    """
-    newsletter_opt_in = models.BooleanField(
-        default=False,
-        help_text="Whether the user signed up for the newsletter",
-    )
-
-    def __str__(self):
-        return f"Donate Form Submission from {self.name} on {self.date_submitted}"
-
-
-class InstrumentRentalRequestSubmission(BaseFormSubmission):
-    """Member portal instrument rental requests. `message` field stores optional notes."""
-
-    STATUS_PENDING = "pending"
-    STATUS_APPROVED = "approved"
-    STATUS_DENIED = "denied"
-    STATUS_RETURNED = "returned"
-    STATUS_CHOICES = [
-        (STATUS_PENDING, "Pending"),
-        (STATUS_APPROVED, "Approved"),
-        (STATUS_DENIED, "Denied"),
-        (STATUS_RETURNED, "Returned"),
-    ]
-
-    member = models.ForeignKey(
-        "blowcomotion.Member",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="rental_requests",
-    )
-    instrument = models.ForeignKey(
-        "blowcomotion.Instrument",
-        on_delete=models.PROTECT,
-        related_name="rental_requests",
-    )
-    second_choice = models.ForeignKey(
-        "blowcomotion.Instrument",
-        null=True,
-        blank=True,
-        on_delete=models.PROTECT,
-        related_name="rental_requests_second_choice",
-    )
-    third_choice = models.ForeignKey(
-        "blowcomotion.Instrument",
-        null=True,
-        blank=True,
-        on_delete=models.PROTECT,
-        related_name="rental_requests_third_choice",
-    )
-    is_waitlist = models.BooleanField(default=False)
-    phone = models.CharField(max_length=255, blank=True, null=True)
-    address = models.CharField(max_length=255, blank=True, null=True)
-    policy_acknowledged = models.BooleanField(default=False)
-    status = models.CharField(
-        max_length=20,
-        choices=STATUS_CHOICES,
-        default=STATUS_PENDING,
-    )
-    patreon_validated = models.BooleanField(
-        null=True,
-        blank=True,
-        help_text="Set automatically via Patreon API at submission time. Active = confirmed; Inactive = not found or inactive; Unknown = API not configured or check failed.",
-    )
-    patreon_pledge_cents = models.PositiveIntegerField(null=True, blank=True, help_text="Monthly pledge amount in cents at time of last Patreon check.")
-    patreon_last_charge_date = models.DateTimeField(null=True, blank=True, help_text="Date of last Patreon charge attempt.")
-    patreon_last_charge_status = models.CharField(max_length=20, null=True, blank=True, help_text="Last Patreon charge status at time of check (e.g. Paid, Declined).")
-    patreon_patron_since = models.DateTimeField(null=True, blank=True, help_text="Date the member started their Patreon pledge.")
-    patreon_lifetime_cents = models.PositiveIntegerField(null=True, blank=True, help_text="Lifetime Patreon support in cents at time of last check.")
-    admin_message = models.TextField(blank=True)
-    assigned_unit = models.ForeignKey(
-        "blowcomotion.LibraryInstrument",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="rental_assignments",
-    )
-    prior_storage_location = models.ForeignKey(
-        "blowcomotion.InstrumentStorageLocation",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="+",
-        help_text="Storage location the instrument was in before being rented out. Restored on return.",
-    )
-
-    def __str__(self):
-        return f"{self.name} — {self.instrument} ({self.status}) on {self.date_submitted:%Y-%m-%d}"
-
-    class Meta:
-        ordering = ["-date_submitted"]
-        verbose_name = "Instrument Rental Request"
-        verbose_name_plural = "Instrument Rental Requests"
